places-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2